+++ packages/ext/wxt.config.ts
@@ -64,7 +64,14 @@ export default defineConfig({
           "default": "Alt+S",
           "mac": "Command+S"
         },
-        "description": "open summary panel in current focus webpage"
+        "description": "__MSG_Commad_Open_Panel_DESC__"
+      },
+      "COMMAND_ADD_SELECTION": {
+        "suggested_key": {
+          "default": "Alt+A",
+          "mac": "Command+A"
+        },
+        "description": "__MSG_add_selection_to_chat__"
       }
     },
     web_accessible_resources: [