+++ CHANGELOG.md
@@ -4,8 +4,10 @@
 1. add adaptation to SPA (Single Page Application), input page content will change when SPA change route.
 2. fix: now chatting directly without first summarizing will send messages with page context.
 3. fix: floating ball y-position using `px` causing it out of view in screen adjustment situations such as changing from vertical to horizontal.
-4. feature: add a context menu item to add selection to chat input.
+4. feature: a context menu item and a command for adding selection to chat input.
 5. optimize prompt item view, add prompt preset view for creating prompt with,  add another two prompt presets, and with langs of `zh-CN` and `zh-TW`.
+6. feature: add clear/reset for clearing all the messages in panel's dialog.
+7. fix: dragging of panel using `px` causing it out of view in screen adjustment situations such as changing from vertical to horizontal.
 
 
 ## [0.1.5] - 2025-02-27